Crate rust_vision

Computer vision algorithms for the ImageLink project.

This crate provides high-performance vision primitives such as corner detection, template matching, and feature extraction, exported for use in the Wolfram Language.

Functions

connected_components_memory

Performs connected components labeling on a binary image.

corners_fast9_memory

Detects corners using the FAST-9 algorithm.

corners_fast12_memory

Detects corners using the FAST-12 algorithm.

distance_transform_memory

Computes the distance transform of a binary image.

evaluate_haar_features_memory

Evaluates all Haar features for a given frame size on an image.

hog_memory

Computes the Histogram of Oriented Gradients (HOG) descriptor.

horizontal_prewitt_memory

horizontal_scharr_memory

Computes the horizontal Scharr operator for an image.

horizontal_sobel_memory

local_binary_pattern_image_memory

Computes the Local Binary Pattern (LBP) image.

match_template_memory

Performs template matching between an image and a template.

prewitt_gradients_memory

Computes Prewitt gradients (magnitude) of an image.

sobel_gradients_memory

Computes Sobel gradients (magnitude) of an image.

suppress_non_maximum_memory

Performs non-maximum suppression on an image.

vertical_prewitt_memory

vertical_scharr_memory

Computes the vertical Scharr operator for an image.

vertical_sobel_memory

vision_get_version

Returns the version string of the rust_vision crate.